Bookmarking bugfix for UTF-8 Everywhere
Using the DEBUG_UTF8_EVERYWHERE setting caught a potential problem
with the updating of bookmarks, if VAL_STRING_AT() operations were
to consider freeing bookmarks as one of its potential policy
choices.  Currently it does not do so--but the fuzzing setting of
periodically dropping bookmarks simulated it, which could have caused
a problem in the Modify_String() code if it updated its idea of the
bookmark without noticing changes made by the operation.

This moves the bookmark extraction to after the operations that can
potentially shuffle bookmarks.