Systemic errors for index-out-of-range errors
There has been a latent potential for crashing or bad semantics with
out-of-bounds indices.  Some of these errors are covered up due to
the fact that series memory isn't generally reclaimed to be smaller.
But also, the out-of-bounds cases are somewhat hard to trigger.

This makes a long-suggested change that the embedded index inside a
series can take on any signed integer value...and that there is more
checking on VAL_INDEX() and VAL_ARRAY_AT() accesses to ensure that
code that cannot meaningfully interpret the out of bounds access does
not do so.

Among the many routines that now error on out of bounds access is
MOLD.  A good answer for what this should do is not currently known,
so until someone presents a motivating use case why it should do
anything in particular it will error as well.

As a first step to mitigating some of the performance concerns that
this change will introduce, this merges together VAL_LEN_AT() with
VAL_ARRAY_AT() to produce VAL_ARRAY_LEN_AT().  By getting the length
at the same time that the pointer is returned, range checking on index
validity only needs to be performed once.  (There will be more places
where this kind of change will be needed.)