VAL_BIN_AT()+VAL_LEN_AT() => VAL_BINARY_SIZE_AT()
This builds on an train of thought about how to reduce the amount of
redundant index validity checking from the previous commit, which
folded together VAL_LEN_AT() and VAL_ARRAY_AT() to get a routine that
returned the length and the data pointer with only one range check on
the index.

Here the same technique is applied to binaries.  It's also a chance
to use the REBSIZ datatype to help call out places that are explicitly
speaking about byte sizes as opposed to the more generic term of
"length".

Rather than converting code in the crypt extension to use the new
function, this switches it over to using the API...as the code had
been written prior to the existence of the rebBytes() extractor.
It still may be the case that read-only locks of values could be
taken through the API to use their bytes directly and avoid a copy,
but starting from rebBytes() is better.