Outdent PARSE code, format comments
The majority of the PARSE code was under a large while() loop, which
added a level of indentation to everything.  Using `goto` to more
explicitly label what a `continue` or `break` is doing has become
a standard way of dealing with this.

This replicates that decision from the stackless branch, and also
brings all the comment formatting in line with the `two spaces off
for line comments" rule.