Eliminate Enter_Native() on INCLUDE_PARAMS macros
With the advent of FRAME!s for debugging, a frame object could be
incarnated for natives to inspect the internal state of that native's
variables.  But being able to change the values in these frames would
be dangerous and could crash the interpreter if a bit pattern the
C code wasn't expecting were used.

To avoid this problem in a performant way, the INCLUDE_PARAMS_OF macros
added in an Enter_Native() call.  This had the undesirable property
that if multiple inclusions occurred (e.g. calling a service routine
which wanted to access the frame) then the Enter_Native() could be
called redundantly.  That was occurring with PARSE in the recent
commits.

This changes to a nearly-as-performant variation that avoids the
problem, by using the IS_NATIVE flag...and lining that flag up with
the same value as SERIES_INFO_HOLD.  It also helps verify that
IS_NATIVE is set correctly by changing the idea of having the params
macro set the hold variable, to testing to make sure it is set.