Eliminate Null_Dispatcher
Originally, `do []` was a null... and so was the result of `func [] []`

This changed to be void instead.  The FUNC code, however, was still
optimizing the case of `func [] []` in an inconsistent way for type
checking so that if you said `func [return: [integer!]] []` it would
think it didn't have to typecheck it.

The null dispatcher needed to be eliminated, and type check exemption
updated to check for void--not null.