Make modal parameters invisibility-sensitive
A typical evaluation can distinguish when the `<end>` of an evaluative
input is reached vs. an evaluative null, by looking at the
CELL_FLAG_UNEVALUATED.  This is something that is not generalized but
is available on frame arguments at the time of frame initialization.

One place it is possible to expose this bit is in modal parameters,
where they could distinguish true invisibility from a null.  e.g.
they can tell @(null) from @() or @(comment "hi").  This presents a
useful opption for building invisibility-aware functions.