Rebind_Values mutability hack
R3-Alpha had several places in the code that did not make copies of
blocks in order to change their binding.  Conceptually, this violates
"read-only-ness" because the data is being changed.  But until the
system is redesigned to more intelligently deal with "virtualness" of
binding, these places have to bypass the mutability check to keep
historical code working.

One place where the MUTABLE_HACK() was needed it was not being used.
This was discovered by running the httpd.reb stress test.