Eliminate `<dequote>`/`<requote>` parameter tags
An experiment from the early implementation of generalized quoting was
a way to automatically remove the quotes from parameters of a function
and apply it to the result.  The feature was mostly guided by wanting
`get first ['x]` to work, since it had previously worked with LIT-WORD!
so it wanted to work with QUOTED! too.

The added complexity and oddness in the evaluator from making this
built in was bad.  Such permissiveness also encouraged letting quotes
go through a lot of routines unchallenged which could make it so that
they could contimatate in a way that would be kind of "dirt" on the
various values in a dialect.

This removes the annotations, and for the moment callers will have to
DEQUOTE or UNQUOTE their arguments to use GET or SET.  Next plan is
to introduce a REQUOTE function that offers the ability at callsites.