Non-Parameter-taking refinements use #
Rebol2 had refinements that were either NONE! if unused, or #[true]
if used.  Red changed this to be #[false] if unused and #[true]
if used.

With a merge of refinements with their arguments in Ren-C, NULL
became the universal sign of "refinement not supplied", that could
be contrasted with any value supplied.  That meant it was possible
to distinguish a BLANK! value from a non-provided refinement.
Hence things like ARRAY/INITIAL which supplied an initial value
for array arguments could consciously distinguish providing a
BLANK! from not having /INITIAL specified at the callsite.

This left a question of what should happen when a refinement had
no arguments, but was only "specified or unspecified".  This is
necessarily different than a refinement that takes a LOGIC!, as
such a refinement would be a tristate: NULL would mean not
provided, and then either #[true] or #[false].

Ideas of using the name of the refinement itself created some
difficulty, because then either the users would be burdened with
having to provide the refinement in that named form -or- their
provision with a LOGIC! would need to be mutated in the frame
at runtime.  This mutation is sketchy.

Also, the concept of using #[true] and NULL would have a similar
sketchiness of allowing users to write #[false] in the argument
but then converting it to NULL at runtime.

This commit solves the problem by choosing `#` as the truthy
state in juxtaposition to NULL.  This helps avoid suggesting that
a #[false] is legal, and requires no canonization.