Labeled Voids (Note; use ~void~ instead of VOID)
This introduces a new concept that makes VOID! act something like a
WORD! that stores a symbol.

Voids begin and end with tilde characters (this means that ~ is no
longer a valid word character).  Voids thus look like:

   ~undefined~
   ~void~
   ~empty~

The label of a void can be retrieved as a WORD! via LABEL OF

   >> label of ~undefined~
   == undefined

To make sure all labeled VOID! can be converted to WORD!, they
cannot contain internal ~ characters.

If a void doesn't have a label (the only kind that does not have
a word conversion), it appears as a single tilde.  The label will
be returned as NULL.

   >> label of ~
   ; null

A void with no label will not be displayed by the console.  But a
conventional function return now produces ~void~, which will be
displayed.  This means functions like HELP have to say `return ~` in
order to suppress the visual.

Voids are still ornery in the sense that they cannot be fetched from
WORD! or PATH! without GET/ANY, and they are not logically true or
false but raise errors there.  But this relaxes the handling of voids
in other places, so they act inert in the evaluator.

The VOID generator function is removed.  While it could be kept and
return the default ~void~, confusions have historically arisen due
to the fact that `type of get/any void` is ACTION! and not VOID!.
When this drawback is combined with the more palatable and general
literal form, then asking people to use that literal form lets them
think about whether there is a meaningful name for their purpose.
If there is no such meaningful name, they have the even more concise
shortcut of just `~` to choose.