Convert | to , usages ... remove | from bootstrap
This moves to full use of the new COMMA! feature, removing the
definition of | as a barrier.

Because | is a distinct datatype (BAR!) in the bootstrap executable,
it cannot be undefined.  Rather than keep | for compatibility, the
few references in the bootstrap scripts are simply removed.

Because barrier behavior is something that can be given to usermode
constructs, the tests formerly for `|` are redefined as being for `||`,
which is a new name for the barrier to help keep that tested.