Remove DEFAULT behavior for CASE/SWITCH
A trick was explored which allowed DEFAULT to become a "skippable
left-quoting enfix operator".  It gained an additional feature of being
able to be used in CASE or SWITCH:

    case [
       1 = 2 [print "Not run"]
       default [print "This would run"]
    ]

The core of the trick was to make this behave exactly like:

    case [
       1 = 2 [print "Not run"]
       (print "This would run")
    ]

It simply noticed when its left-hand side wasn't a SET-WORD!, and
would decay to this behavior.  The behavior of CASE and SWITCH to let
the last condition value "fall out" if there was no associated code
block to run allowed it to work.

HOWEVER in the era of predicates, the issue becomes more complex.  As
a simple example:

     case .greater? [
         1 2 [print "Not run"]
         default [print "The intent would be that this run"]
     ]

The mechanim collapses.  It also points to a question about fallout
in general...showing it cannot let the *input* to the predicate fall
out, but only the *output*...which is less useful.

It also had a problem in its interaction with CASE/ALL, because it
meant the DEFAULT would be run as well...even if no other branches ran.

While exploring this feature was a good exercise, it stands in the way
of progress on predicate development.  And it's covered more rigorously
by ELSE.  Hence this removes this aspect of DEFAULT (it still works
when the left is a SET-WORD! for defaulting variables).