MAKE FRAME! defaults values to ~undefined~
In the initial development of MAKE FRAME! and specialization,
the VOID! type did not exist.  So it seemed clear that NULL should
be used as the starting state for a cleanly created frame.

Now that time has passed, VOID! seems the more obvious choice.
This means that frame specialization can treat `~undefined~`
states as those that should be fulfilled at the callsite...so you
can easily specialize with NULL.  This is more common than
wanting to specialize with `~undefined~`.

As an added benefit of labeled voids, this means there is only
*one* VOID! state that is "difficult" to specialize with.  (It
should still be possible to specialize with it by using the
feature of hiding fields in FRAME!s, TBD)