Introduce REFRAMER and REQUOTE
ENCLOSE is a tool for creating a function that can manipulate the
FRAME! of a function call before it runs, and then act on the
result of the frame execution (if it chooses to execute it).

This introduces the REFRAMER, which is related, but that does not
connect the manipulator to a particular function ahead of time.
Instead, the manipulator gets a frame that is created for the
function invocation immediately after it.

A motivating example is included, called REQUOTE.  This will
take the quote levels off the first argument to a function, then
invoke it, then add the quote levels back to the result (if it
is not a NULL)

     requote: reframer func [f [frame!]] [
         p: first words of f
         num-quotes: quotes of f/(p)

         f/(p): dequote f/(p)

         return quote/depth do f num-quotes
     ]

     >> item: first ['''[a b c]]

     >> requote append item <d>  ; append doesn't accept QUOTED! items
     == '''[a b c <d>]   ; munging frame and result makes it seem to