Differentiate archetypal frames via label node
This makes sure that all frames which are not running have at least
a difference from the running frames by having either a REBSTR* node
or a nullptr where the phase would be.

What can be accomplished with this is that an archetypal frame can
know not to display locals, while a non-archetypal frame can show
those locals.

e.g. this represents two different views of a function for the same
frame, based on the phase stored in the FRAME! value

    >> foo: func [x <local> y] [return binding of 'x]

    >> f: make frame! :foo
    == make frame! [
        x: ~undefined~
    ]

    >> foo
    == make frame! [
        return: 'make action! [[@value /vanishable] [...]]
        x: 10
        y: ~undefined~
    ]