Eliminate REB_TS_HIDDEN and REB_TS_UNBINDABLE
This gets rid of the two type flags for the hidden and unbindable
states.  Instead it uses parameter classes (acquisition modes) of
LOCAL and SEALED.

In the near term, this helps with opening up datatypes in the 64-bit
limit of typeset flags.  But more generally, it reduces the
combinatorics that need to be worried about in terms of keeping the
flags in sync.