Beginning Work on FRAME! Information Hiding
This is a rough start to the idea that frames can have multiple
instances of the same named key...where which phase is in effect
dictates which version of the key is available.

It means that functions like AUGMENT can add new parameters that have
the same name as locals in the inner layers of functions.

Implementation-wise, this requires routines that formerly just took
a REBCTX* to take a full ANY-CONTEXT! value, so as to extract the
phase of frames and use that to inform which keys are active.