Limit ADAPT/ENCLOSE/ETC to ACTION!, no WORD!/PATH!
When functions like ADAPT were being first designed, all ACTION!s
were fully anonymous.  So there was no label known for them.

To get around this, an alternative was offered to passing in an
ACTION!.  Instead, you could pass in a WORD! or PATH!, that would
be processed in such a way as to preserve the name:

    >> specialize :append [value: 10]
    ; ^-- this SPECIALIZE couldn't know the name "append"

    >> specialize 'append [value: 10]
    ; ^-- this SPECIALIZE has the WORD! in hand, does the GET
    ; hence it can also know the name

Yet new features have made this obsolete!  ACTION!s hold labels that
come from them being obtained through GET-WORD!s or GET-PATH!s.  These
labels live in cells and are trickily moved around to persist with
the actions.

This means the confusing type signature and quoting mechanics no longer
make sense.  This changes to only allow ACTION!.