DO of GROUP! forms is invisible
Ren-C attempted to limit the behavior of DO such that it would have
more "interesting" behaviors (rather than say `do` of a TAG! would
be the value of the tag, due to not having any better ideas).  So it
made things that were conventionally allowed to pass through errors
since they provided no new behavior.

This led to a question of whether DO of a GROUP! should be the same
as a BLOCK!, or if there was an interesting subtlety.  Eventually a
subtlety emerged that is somewhat interesting: that DO of a GROUP!
could vaporize on empty blocks instead of returning void:

    >> 1 + 2 do []
    ; void result

    >> 1 + 2 do '()
    == 3

This also works for comments or other completely invisible contents:

    >> 1 + 2 do '(comment "hi")
    == 3