Make AND, OR, XOR take GROUP! on right hand side
The idea of using AND/OR/XOR as conditional to match the expectations
of other languages goes hand-in-hand with another concept that people
would expect: short-circuit behavior.

In order for a right hand expression to be conditionally executed,
it would have to be in a GROUP! or a BLOCK!.  Using GROUP! would
look better, but would require quoting...which was considered to be
a "fringe" parameter convention concept.  This meant the early
development of conditional AND used a BLOCK! on the right hand side
for conditional execution, but permitted GROUP! if both sides were
to be executed with no short-circuiting.

Philosophy evolved to where quoted parameters were seen as a more
mainstream approach to engineering the calling conventions to get
the "dialect" that users wanted.  This led to GROUP!s being quoted
by branching, but only executed to build branches when the condition
matched.  With this tolerated, it's not much of a stretch to use plain
GROUP! for short circuiting branches...which is quite readable.

    >> false and (print "right" true)
    == #[false]

    >> true and (print "right" true)
    right
    == #[true]

GET-GROUP! can then be used to subvert this, running the right hand
side regardless:

    >> false and :(print "right" true)
    right
    == #[false]

Shorthands using @word and @pa/th for the right hand side are also
provided.