Switch FAIL to @word for blame, word is ID
The ability of FAIL to use a skippable parameter to indicate what
site to blame was one of the first purposes for @word, because the
@ seemed like a way of "pointing at" something.

When the @ became more used for concepts of "literally, this", the
pointing concept didn't make as much sense.  `@x` was frequently being
used to mean "fetch the contents of x, and process them literally", or

That made it a bit less of a slam dunk, so it was backed off to the
previous choice of `fail 'x` to indicate "fail and implicate the
X parameter".

With time going by, it seems like allowing FAIL this usage of @x is
in line with the kind of choice a dialected special function would
make.  And it frees up errors to be raised by a WORD! id, where the
ID could more generally be a PATH! or a URL!.

So this does some improvements to FAIL and switches back to the @x
form for blame.  Also adds a test file, since none existed.