%% as WORD! (introduced for MAKE PATH! operator)
This introduces a "file-suggesting" operator of two percent signs.
Two is allowed, while one is not, because: `%/foo` would be ambiguous
as being either a PATH! or a FILE!.  It is worth it to sacrifice the
`%%` FILE! representation for an operator that can be as useful as it
can be as a WORD!, however.  (The FILE! can be done as %"%")