Patch TCC encapping for MAKE-DIR
This was only being run in deployment.  It was also running before the
executable got a platform-specific name (making it harder to find if
it were actually being encapped).

This avoids using a stray version for the act of encapping, instead
using the built r3.  That means Windows builds don't get the encap...
which isn't terribly relevant since encapping is not being performed
at this time.