Add a Rust-like "option" and "unwrap" facility
The ability to annotate when a value is "optional" is something that
has been embraced by the mainstream, e.g. in Rust.  Ren-C had
informally annotated parameters that were optional with `opt_` in
their names, but this takes a further step to make it so that pointers
which use `nullptr` as an optional state can be enforced at
compile time.

The feature is only active in checked builds with DEBUG_CHECK_OPTIONALS
and is purely commentary otherwise.