Feeds use a singular array for current source
The concept behind this change is that a feed would be able to accept
a linked list of "splices" from which to draw input.  Each splice
would be a singular array.

Feeds incorporate allocation space for one such array by default, but
if a splice is made then the content of that array is moved out to
a new node to make room for the splice.