Simplify feed fetching and lookback mechanics
The code for feeds and their ability to operate variadically has grown
in a very "organic" fashion.

One concern that the code has is regarding the pointer stability of
values that are "behind" the current one that may need to be "looked
back at".  This can be an unstable situation because some values are
synthesized into a feed's "fetched" cell, and retrieving the next
value can wind up overwriting those bits.

Since that mechanic has settled somewhat, this establishes a simpler
rule that all lookbacks are considered stable unless the fetched cell
is used.