Add MACRO function generator
This creates a function category that splices content into the
execution stream:

    >> m: macro [x] [return [append x first]]

    >> m [1 2 3] [d e f]
    == [1 2 3 d]

While expedient, this does have drawbacks.  For instance: while the
function appears to take one parameter, in practice it will act as
if it takes two.  It will not have the same composability with things
like SPECIALIZE or ADAPT that a formally specified function would.

It is introduced largely as a test of a new generalized splicing
ability, which is intended for use with predicates.