Switch to using macro mechanism for predicates
This switches over predicates to not necessarily need to generate a
function, but splice raw material into the evaluative stream, which
should be significantly more efficient.

If a predicate is a GROUP!, then it has an exemption to be able to
run the generated product vs. just dropping a function value inertly
in the stream.