Add predicates to ANY and ALL, remove NONE
This brings the predicate feature to ANY and ALL, so they can be used
with functions that determine what constitutes passing the "test"
(besides merely being truthy).

This makes the NONE function redundant with `ALL .NOT` so its
implementation is removed.  It may be that NONE is still useful, but
it would be done as a specialization vs. something that would take
a predicate in its own right.