Add "raw cell" with no C++11 protections
Value cells were already divided into REBVAL, RELVAL, and REBCEL...
each performing a distinct purpose.  However, all forms were not
able to be assigned or copied.  This created complexity in lower
level functions that want to do things like memcpy() a series which
contains a cell.  (Which should not be done usually, but these
privileged routines know what they are doing).

This creates a special REBRAW cell class for this narrow purpose,
that most of the system won't be aware of.