Make BLANK!-headed TUPLE! a local in function specs
Previously locals were represented as individual items in parameter
lists as SET-WORD!s.  This changes that to use blank headed tuples.
So `.foo` instead of `foo:`

The goal is to retake SET-WORD!s to use exclusively for return values
(and RETURN: may be an exception, as a local that *represents* the
main return, but holds the definitional return function).