Change output parameters to SET-WORD! in specs
This makes the sensible-seeming change to the function spec dialect
that SET-WORD! represents return parameters.

It replaces the `<output>` annotation with `<in-out>`, that is planned
for use at a lter time.  However, to make the SET-WORD!s for outputs
easier to find for the moment, the parameter descriptions are changed
to include the string `<output>`.

This requires changes in a subsequent commit to function, but those
changes are kept separate for clarity.