Systemize ARG_MARKED_CHECKED for parameter hiding
The relationship between ARG_MARKED_CHECKED and hidden parameters has
been a bit nebulous, and uncertain whether every source of frame
values was keeping it in sync.

One feature that was of particular interest was the ability to mark
a parameter as being intentionally specialized as ~undefined~.  The
idea for conveying this state for such parameters was to hide it.
But hiding parameters by manipulating their properties in a keylist
would affect all frames and invocations tied to a function unless
the keylist was copied.

This systemizes ARG_MARKED_CHECKED a bit more so it's kind of more
like VAR_MARKED_HIDDEN, and changes parameter hidden state to come
from this bit.  It's also helpful to not destroy the parameter
information (like REB_P_OUTPUT).  Using a flag that's not in the
paramlist accomplishes that.

It also means that SPECIALIZE does not need to tweak the parameter
convention bits in the paramlist it is specializing.  This opens up
the door to potential compressions in representation where the
paramlist of the unspecialized function gets reused.