Sub-second date scanning, better scanner comment
The switch of dot into a delimiter lexical class threw off the
loading of dates with time components that had decimal numbers for
the seconds.

This addresses that.  There was also some dead code that obfuscated
the DECIMAL! scanning exception to TUPLE!.  So that code is removed,
and the comment explaining it updated some.