Separate CHAIN Dispatcher Mechanic from Evaluator
The CHAIN function was one of the earliest function combining
generators.  When it was written, frames were considerably less
flexible; so the only mechanic to reuse one was to change the
dispatcher and relinquish control to the evaluator.  Since CHAIN did
not want to fully relinquish control, the data stack was used to push
actions which the evaluator noticed after processing and would run
on the outputs.

The introduction of ENCLOSE and other operations made frames more
versatile at being unplugged and replugged freely.  The rearrangement
capabilities are converging to be more extreme in the "stackless"
branch of the interpreter.

Hence a more sensible idea for CHAIN is to have it pass off the built
frame contents to a new REBFRM* which it executes while keeping the
dispatcher running the chain linked into the stack.  A key benefit
of this is that the evaluator does not have to be explicitly involved
with the mechanic.  Another is that error messages can be clearer
as errors resulting from a step in the chain will have that step's
name on the stack frame...as well as the chain itself above, so both
name will be seen.

A hard-to-understand mechanic in the formative RESKINNED primitive was
making use of the CHAIN's hook into the evaluator.  This removes that
hook and just has RESKINNED do the same thing as CHAIN when it wants
to manipulate the return result of a function it built a frame for;
it "downshifts" the frame contents into a new REBFRM*, which it runs
while it remains linked into the stack, and then post-processes the
result of executing that.