Date Comparison Fix
Attempting to simplify the comparison interface moved the "strict"
flag into a clear boolean argument to the comparison operations, and
forced all those comparisons to return -1, 0, or 1.

The `>` and `<` comparison operators--which previously did not have
strict forms--would ask for the tristate with the `strict` flag on.

Since the strict handling for dates was only written to handle the
question of equality or inequality, it took a shortcut to answer
that particular question without looking at the dates.  This meant
date comparison with > and < became generally broken.  However, there
wasn't anything in the tests to catch the problem.