Remove scratch cell parameter to Do_Branch_Throws
Because some forms of branch evaluation might need to do a temporary
evaluation, it was taking a parameter from the callsite for that cell.
Ultimately this should be paid for out of the sunk cost of making a
frame and feed (the stackless branch did so).  The code isn't factored
to make that easy at the moment, but this improves the callsite
experience just by adding an extra temp cell inside the routine.