Eliminate ACT_UNDERLYING(), meta on details
Previously there was separate storage of the paramlist and a
"specialty"...which was either the array of the paramlist or the array
of an exemplar.  Since the paramlist can be found from the exemplar,
these became redundant.  As a consequence, the ACT_META() could be
moved onto the details array into the freed space...thus making it
possible to completely share paramlists.

As part of the rearrangement, the generated `meta` object is now
returned as a separate return value from the paramlist creation process,
making it easier to put on something besides the paramlist.  (This is
needed now that the paramlist can be shared, so the meta had to move
onto the details node.)

Additionally, the ACT_UNDERLYING() concept is replaced by the idea
that paramlists follow the same form as keylists, using an ancestor
pointer.  This pointer is to a paramlist of the same or shorter
length, and frames built for a derived paramlist are expected to be
compatible with an action with an ancestor paramlist.