Clarify some hacks related to action flags
In order to share common code for processing parameter lists and
digesting information out of them, some edge case features of that
parameter flag analysis were snuck onto series flags instead of
adding an extra parameter.

This got increasingly sketchy over time; and it wasn't clear which
cases were handling the flags vs. not.  This bites the bullet and
makes the parameterization of these flags explicit as part of the
parameter list analysis...getting them out of the more hidden locations.

Also, since the action flags are now cached on the details array,
the PARAMLIST_FLAGs are renamed as DETAILS_FLAGs.  The exception is
for the paramlist flag for HAS_RETURN, which is a can of worms needing
more in-depth treatment.

This cleans up several transitional hacks in the switch from paramlist
to details for action identity.